refactor: 패키지 구조 이동 및 코드 정리
