+++ src/main/kotlin/racing/domain/Game.kt
@@ -1,11 +1,10 @@
-package racing
+package racing.domain
 
-import racing.domain.Car
-import racing.domain.RandomStrategy
+import racing.domain.strategy.RandomMovingStrategy
 
 data class Game(val team: Int, val round: Int) {
     companion object {
-        private val RANDOM_STRATEGY = RandomStrategy()
+        private val RANDOM_STRATEGY = RandomMovingStrategy()
     }
 
     val cars = List(team) { Car(RANDOM_STRATEGY) }
