--- src/main/kotlin/racing/domain/MovingStrategy.kt
@@ -1,5 +1,0 @@
-package racing.domain
-
-interface MovingStrategy {
-    fun move(): Boolean = true
-}
