+++ src/main/kotlin/racing/domain/strategy/MovingStrategy.kt
@@ -1,0 +1,5 @@
+package racing.domain.strategy
+
+interface MovingStrategy {
+    fun move(): Boolean
+}
