+++ src/main/kotlin/racing/domain/strategy/RandomMovingStrategy.kt
@@ -1,8 +1,8 @@
-package racing.domain
+package racing.domain.strategy
 
 import java.util.Random
 
-class RandomStrategy : MovingStrategy {
+class RandomMovingStrategy : MovingStrategy {
     companion object {
         const val MOVING_POINT = 4
         const val MOVING_RANGE_MAX = 10
